Move to_dict, between, more constructors
